backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
# O projeto usa bcrypt direto (com fallback), nao depende de passlib em runtime.
bcrypt>=4.2.0
python-multipart==0.0.9
pydantic>=2.10.0,<3
pydantic-settings>=2.6.0,<3
httpx==0.27.0
pythonnet==3.0.5
requests==2.31.0
python-dotenv==1.0.1
aiofiles==23.2.1
